transfer highcharts to components
